% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{consulta_enaho}
\alias{consulta_enaho}
\title{Encuesta Nacional de Hogares versión anual - Metodología ACTUALIZADA}
\usage{
consulta_enaho(
  periodo,
  codigo_modulo,
  base,
  guardar = FALSE,
  ruta = "",
  codificacion = NULL
)
}
\arguments{
\item{consulta_enaho}{Consulta a ENAHO de cada año en metodología actualizada}
}
\value{
Base del módulo de interés de la ENAHO
}
\description{
Consulta las bases de datos de la ENAHO de forma directa
}
\examples{
temp1 <- consulta_enaho(periodo = 2020, codigo_modulo = 1, base = "enaho01-100-2022")
}
