% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{consulta_enapres}
\alias{consulta_enapres}
\title{Encuesta Nacional de Programas Presupuestales}
\usage{
consulta_enapres(
  periodo,
  codigo_modulo,
  base,
  guardar = F,
  ruta = "",
  codificacion = NULL
)
}
\arguments{
\item{consulta_enapres}{Consulta a módulo de interés de la ENAPRES según año}
}
\value{
Módulo de la ENAPRES consultado
}
\description{
Consulta directa a la ENAPRES del INEI
}
\examples{
temp1 <- consulta_enapres(periodo = 2019, codigo_modulo = 1492, base = "CAP_300_URBANO_RURAL_5")
}
