% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{consulta_endes}
\alias{consulta_endes}
\title{Encuesta Demográfica y de Salud Familiar - ENDES}
\usage{
consulta_endes(
  periodo,
  codigo_modulo,
  base,
  guardar = FALSE,
  ruta = "",
  codificacion = NULL
)
}
\arguments{
\item{periodo}{Año de consulta}

\item{codigo_modulo}{ID del módulo (Dos dígitos para la ENDES)}

\item{base}{Nombre de la base a consultar}
}
\value{
Base del módulo de interés de la ENDES
}
\description{
Consulta las bases de datos de la ENDES de forma directa
}
\examples{
temp1 <- consulta_endes(periodo = 2004, codigo_modulo = 64, base = "RECH0")
}
