% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{inei_tabla}
\alias{inei_tabla}
\title{Tablas en formato INEI}
\usage{
inei_tabla(
  wb,
  sheet,
  data,
  cuadro,
  titulo,
  subtitulo,
  fuente = NULL,
  nota = NULL,
  formato = "0\%"
)
}
\description{
Tablas en formato INEI
}
